extract isRgbaFloat
